+++ app/JsonApi/Partner/V1/Employees/EmployeeRequest.php
@@ -5,39 +5,45 @@ namespace App\JsonApi\Partner\V1\Employees;
 use Illuminate\Validation\Rule;
 use LaravelJsonApi\Laravel\Http\Requests\ResourceRequest;
 use LaravelJsonApi\Validation\Rule as JsonApiRule;
+use App\JsonApi\Admin\V1\PartnerUsers\PartnerUserRequest;
 
-class EmployeeRequest extends ResourceRequest
+
+class EmployeeRequest extends PartnerUserRequest
 {
 
     protected function prepareForValidation()
     {
 
-        if($this->isMethod("DELETE")) return;
-
-        $data = $this->data;
-        $data['attributes']['phone'] = preg_replace('/[^0-9]/', '', $this->data['attributes']['phone']);
-
-        $this->getInputSource()->replace(['data' => $data]);
+        parent::prepareForValidation();
     }
+
+
     /**
      * Get the validation rules for the resource.
+     * Не сработало через наследование, т.к. нужно передать базовую модель '$partner_user = $this->model()->toBase();'
      *
      * @return array
      */
     public function rules(): array
     {
-        $company_user = $this->model();
+        $partner_user = $this->model()->toBase();
 
         $unique = Rule::unique('partner_user');
 
-        if($company_user) {
-            $unique = $unique->ignore($company_user);
+        if($partner_user) {
+            $unique = $unique->ignore($partner_user);
         }
 
-        return [
+        $rules =  [
+            //'partnerId' => 'required|exists:partners,id',
             'phone' => 'required|digits_between:3,15|' . $unique,
-            "status" => 'boolean'
+            'status' => 'boolean'
         ];
+        if($this->isMethod('PATCH')){
+            unset($rules['phone']); // убрать проверку на номер телефона при редактировании пользователя
+            //unset($rules['partnerId']); // убрать проверку на companyId, т.к. компанию после назначения уже поменять нельзя
+        }
+        return $rules;
     }
 
 }